requirements.txt
asgiref==3.7.2
Django==5.0.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
pillow==10.2.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
